Adding props
